Implemented IEquatable, ICloneable interfaces to MapObjects and Vector2. Changed key inputs to enum. Removed all object diameters introducting radius. Fully removed floats to doubles. Partially working movements (one test not passing). Fixed generating new player's id.

+++ GameTests/PhysicTests.cs
@@ -15,7 +15,7 @@ namespace GameTests
         {
             // Arrage
             var vector = new Vector2(10, 0);
-            var angle = (float)Math.PI / 4;
+            var angle = Math.PI / 4;
 
             // Act
             var rotatedVector = Physic.RotateVector(vector, angle);
@@ -31,7 +31,7 @@ namespace GameTests
         {
             // Arrage
             var vector = new Vector2(0, -3);
-            var angle = (float)Math.PI / -2;
+            var angle = Math.PI / -2;
 
             // Act
             var rotatedVector = Physic.RotateVector(vector, angle);
@@ -131,7 +131,7 @@ namespace GameTests
             var parallelVector = Physic.GetParallelVectorToNormal(vector, normalvector);
 
             // Assert
-            var expectedVector = Vector2.Normalize(new Vector2(-1, 1)) * ((vector.Length() * (float)Math.Sqrt(2)) / 2);
+            var expectedVector = Vector2.Normalize(new Vector2(-1, 1)) * ((vector.Length() * Math.Sqrt(2)) / 2);
             Assert.AreEqual(parallelVector.X, expectedVector.X, 0.0001);
             Assert.AreEqual(parallelVector.Y, expectedVector.Y, 0.0001);
         }
@@ -148,7 +148,7 @@ namespace GameTests
             var parallelVector = Physic.GetLeftParallelVectorToIntersectionNormal(movementVector, intersectionDistance, intersectionNormal);
 
             // Assert
-            var expectedVector = Vector2.Normalize(new Vector2(-1, 1)) * (((movementVector.Length() - intersectionDistance) * (float)Math.Sqrt(2)) / 2);
+            var expectedVector = Vector2.Normalize(new Vector2(-1, 1)) * (((movementVector.Length() - intersectionDistance) * Math.Sqrt(2)) / 2);
             Assert.AreEqual(parallelVector.X, expectedVector.X, 0.0001);
             Assert.AreEqual(parallelVector.Y, expectedVector.Y, 0.0001);
         }
@@ -159,14 +159,14 @@ namespace GameTests
         {
             // Arrage
             var movementVector = new Vector2(-3, 3);
-            var intersectionDistance = (float)Math.Sqrt(2);
+            var intersectionDistance = Math.Sqrt(2);
             var intersectionNormal = new Vector2(-1, 0);
 
             // Act
             var parallelVector = Physic.GetLeftParallelVectorToIntersectionNormal(movementVector, intersectionDistance, intersectionNormal);
 
             // Assert
-            var expectedVector = Vector2.Normalize(new Vector2(0, 1)) * (((movementVector.Length() - intersectionDistance) * (float)Math.Sqrt(2)) / 2);
+            var expectedVector = Vector2.Normalize(new Vector2(0, 1)) * (((movementVector.Length() - intersectionDistance) * Math.Sqrt(2)) / 2);
             Assert.AreEqual(parallelVector.X, expectedVector.X, 0.0001);
             Assert.AreEqual(parallelVector.Y, expectedVector.Y, 0.0001);
         }
@@ -175,7 +175,7 @@ namespace GameTests
         public void Physic_ShouldCalculateVectorFromAngle()
         {
             // Arrange
-            var angle = (float)Math.PI * 3 / 2;
+            var angle = Math.PI * 3 / 2;
 
             // Act
             var vector = Physic.AngleToVector(angle);
@@ -196,7 +196,7 @@ namespace GameTests
             var angle = Physic.VectorToAngle(vector);
 
             // Assert
-            var expectedAngle = (float)Math.PI * -1 / 4;
+            var expectedAngle = Math.PI * -1 / 4;
             Assert.AreEqual(expectedAngle, angle, 0.0001);
         }
     }
